Arreglamos hard fault

+++ TPFworkspace/sensores/workspace/sensoresMedicos/source/bluetooth_com.h
@@ -15,9 +15,9 @@ typedef struct{
 	int32_t sp02;
 	uint16_t temp;
 	int32_t ecg_samples[20];
-	uint8_t n_samples_ecg;
+	uint16_t n_samples_ecg;
 	int32_t ox_samples[5];
-	uint8_t n_samples_ppg;
+	uint16_t n_samples_ppg;
 }data_BT_t;
